Refactored structure using sealed class Implement Response.map in terms of flatMap
