+++ src/main/kotlin/com/fpinbo/kall/response/ResponseFunctorExtensions.kt
@@ -1,7 +1,6 @@
 package com.fpinbo.kall.response
 
 fun <A, T> Response<A>.fold(fe: (Response.Error<A>) -> T, fs: (Response.Success<A>) -> T): T {
-
     return when (this) {
         is Response.Error -> fe(this)
         is Response.Success -> fs(this)
@@ -9,8 +8,5 @@ fun <A, T> Response<A>.fold(fe: (Response.Error<A>) -> T, fs: (Response.Success<
 }
 
 fun <A, B> Response<A>.map(f: (A) -> B): Response<B> {
-    return fold(
-        { Response.Error(it.errorBody, code, headers, message) },
-        { Response.Success(f(it.body), code, headers, message) }
-    )
+    return flatMap({ Response.just(f(it)) })
 }
