+++ src/main/kotlin/com/fpinbo/kall/response/ResponseMonadExtensions.kt
@@ -1,7 +1,5 @@
 package com.fpinbo.kall.response
 
 fun <A, B> Response<A>.flatMap(f: (A) -> Response<B>): Response<B> =
-        when (this) {
-            is Response.Error -> Response.Error(errorBody, code, headers, message)
-            is Response.Success -> f(body)
-        }
+        fold({ Response.Error(it.errorBody, code, headers, message) }, { f(it.body) })
+
